# app/api/openapi.yaml
openapi: 3.0.3
info:
  title: API-2-Tool
  version: 1.0.0
  description: |
    API para cadastro de definicoes de tools, listagem, consulta por ID e execucao de chamadas
    upstream com retorno normalizado em function_call_output.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /tool:
    post:
      tags: [Tools]
      summary: Cadastra uma definicao de tool
      operationId: createTool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolDefinitionInput'
      responses:
        '201':
          description: Tool cadastrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDefinitionResponse'
        '400':
          description: Requisicao invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tool/definitions:
    get:
      tags: [Tools]
      summary: Lista todas as definicoes de tools
      operationId: listToolDefinitions
      responses:
        '200':
          description: Lista de tools
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDefinitionsListResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: definition not found
      required: [error]
    ToolDefinitionInput:
      type: object
      properties:
        name:
          type: string
          example: viacep_consultar_cep
        description:
          type: string
          example: Consulta dados de CEP no ViaCEP.
        method:
          type: string
          example: GET
        url:
          type: string
          format: uri
          example: https://viacep.com.br/ws/01001000/json/
        headers:
          type: object
          additionalProperties:
            type: string
          example:
            Authorization: Bearer token
        parameters:
          type: object
          additionalProperties: true
          example:
            type: object
            properties:
              cep:
                type: string
        strict:
          type: boolean
          default: false
      required: [name, method, url]
...
    ToolFunction:
      type: object
      properties:
        name:
          type: string
          example: viacep_consultar_cep
        description:
          type: string
          example: Consulta dados de CEP no ViaCEP.
        parameters:
          type: object
          additionalProperties: true
        strict:
          type: boolean
      required: [name, parameters]
    ToolFunctionDefinition:
      type: object
      properties:
        type:
          type: string
          example: function
        function:
          $ref: '#/components/schemas/ToolFunction'
      required: [type, function]
    UpstreamDefinition:
      type: object
      properties:
        method:
          type: string
          example: GET
        url:
          type: string
          format: uri
          example: https://viacep.com.br/ws/01001000/json/
        headers:
          type: object
          additionalProperties:
            type: string
      required: [method, url]
    ToolDefinitionResponse:
      type: object
      properties:
        id:
          type: string
          example: tool_ba1cb19e-238a-42cb-95d1-2c4a7bf3f4fc
        active:
          type: boolean
          example: true
        tool:
          $ref: '#/components/schemas/ToolFunctionDefinition'
        upstream:
          $ref: '#/components/schemas/UpstreamDefinition'
      required: [id, active, tool, upstream]
    ToolDefinitionsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ToolDefinitionResponse'
      required: [items]
...